move notify event to client
